Improve buildUrl function
The buildURL function was expecting either an array or an object for pathParams. Depending on what it receives it generates a path URL or a query URL. That was not optimal in some situations.

I’ve changed the signature of the function to expect pathParams - array and queryParams - object. This way we can generate a url with both a path and query.

This is helpful in situations where users of the library would like to generate links to user account, contact or payment request but they also need to have some additional query params.

In the past the payment createRequest function was generating different paths depending on whether one provides a subject (it was difficult to match such paths and created a lot of overhead). Now the path is always the same:
/network/:network_address/:user_address/:amount if the user needs additional params to that url they are added as query params ?subject=1&note=note etc

fixes #336