Remove transactionId for transactionHash
We remove the transactionId and replace it by transactionHash.
The transactionId is not truely removed yet, for backwards
compatibility.

To in future not always have failing e2e tests when we add a field to
events, I replaced have.keys with include.all.keys, which only checks
that the keys are a superset of the necessary keys.